feat: fixing lombok, refactoring security to fix /login 403 error, creating AutenticacaoService and general fixes
