--- src/main/java/med/voll/api/controller/AutenticacaoController.java
@@ -1,27 +1,0 @@
-package med.voll.api.controller;
-
-import jakarta.validation.Valid;
-import med.voll.api.domain.usuario.DadosAutenticacao;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/login")
-public class AutenticacaoController {
-
-    private AuthenticationManager manager;
-
-    public AutenticacaoController(AuthenticationManager manager) {
-        this.manager = manager;
-    }
-
-    @PostMapping
-    public ResponseEntity<?> efetuarLogin(@RequestBody @Valid DadosAutenticacao dados) {
-        var token = new UsernamePasswordAuthenticationToken(dados.login(), dados.senha());
-        var authentication = manager.authenticate(token);
-
-        return ResponseEntity.ok().build();
-    }
-}
