--- src/main/java/med/voll/api/controller/MedicoController.java
@@ -1,65 +1,0 @@
-package med.voll.api.controller;
-
-import jakarta.transaction.Transactional;
-import jakarta.validation.Valid;
-import med.voll.api.domain.medico.*;
-import med.voll.api.domain.medico.*;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.sql.init.SqlDataSourceScriptDatabaseInitializer;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.util.UriComponentsBuilder;
-
-@RestController
-@RequestMapping("medicos")
-public class MedicoController {
-    @Autowired
-    private MedicoRepository repository;
-    private SqlDataSourceScriptDatabaseInitializer dataSourceScriptDatabaseInitializer;
-
-    @PostMapping
-    @Transactional
-    public ResponseEntity cadastrar(@RequestBody @Valid DadosCadastroMedico dados, UriComponentsBuilder uriBuilder) {
-        var medico = new Medico(dados);
-        repository.save(medico);
-        var uri = uriBuilder.path("/medicos/{id}").buildAndExpand(medico.getId()).toUri();
-        return ResponseEntity.created(uri).body(new DadosDetalhamentoMedico(medico));
-
-    }
-
-    @GetMapping
-    public ResponseEntity<Page<DadosListagemMedico>> listar(@PageableDefault(size = 10, sort = {"nome"}) Pageable paginacao) {
-        var page = repository.findAllByAtivoTrue(paginacao).map(DadosListagemMedico::new);
-        return ResponseEntity.ok(page);
-
-    }
-
-    @PutMapping
-    @Transactional
-    public ResponseEntity atualizar(@RequestBody @Valid DadosAtualizacaoMedico dados) {
-        var medico = repository.getReferenceById(dados.id());
-        medico.atualizarInformacoes(dados);
-        return ResponseEntity.ok(new DadosDetalhamentoMedico(medico));
-    }
-
-    @DeleteMapping("/{id}")
-    @Transactional
-    public ResponseEntity deleter(@PathVariable Long id) {
-        var medico = repository.getReferenceById(id);
-        medico.desativar();
-        return ResponseEntity.noContent().build();
-
-    }
-
-    @GetMapping("/{id}")
-
-    public ResponseEntity detalhar(@PathVariable Long id) {
-        var medico = repository.getReferenceById(id);
-        return ResponseEntity.ok(new DadosDetalhamentoMedico(medico));
-
-
-    }
-}
