--- src/main/java/med/voll/api/controller/PacienteController.java
@@ -1,57 +1,0 @@
-package med.voll.api.controller;
-
-import jakarta.validation.Valid;
-import med.voll.api.domain.paciente.*;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("pacientes")
-public class PacienteController {
-
-    @Autowired
-    private PacienteRepository repository;
-
-    @PostMapping
-    @Transactional
-    public ResponseEntity<Paciente> cadastrar(@RequestBody @Valid DadosCadastroPaciente dados) {
-        Paciente paciente = new Paciente(dados);
-        repository.save(paciente);
-        return ResponseEntity.ok(paciente);
-    }
-
-    @GetMapping
-    public ResponseEntity<Page<DadosListagemPaciente>> listar(@PageableDefault(size = 10, sort = {"nome"}) Pageable paginacao) {
-        var page = repository.findAllByAtivoTrue(paginacao).map(DadosListagemPaciente::new);
-        return ResponseEntity.ok(page);
-    }
-
-    @PutMapping
-    @Transactional
-    public ResponseEntity<Paciente> atualizar(@RequestBody @Valid DadosAtualizacaoPaciente dados) {
-        var paciente = repository.getReferenceById(dados.id());
-        paciente.atualizarInformacoes(dados);
-        return ResponseEntity.ok(paciente);
-    }
-
-    @DeleteMapping("/{id}")
-    @Transactional
-    public ResponseEntity<Void> excluir(@PathVariable Long id) {
-        var paciente = repository.getReferenceById(id);
-        paciente.excluir();
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity detalhar(@PathVariable Long id) {
-        var paciente = repository.getReferenceById(id);
-        return ResponseEntity.ok(new DadosDetalhamentoPaciente(paciente));
-
-
-    }
-}
