+++ src/main/java/med/voll/api/domain/entities/endereco/DadosEndereco.java
@@ -1,0 +1,27 @@
+package med.voll.api.domain.entities.endereco;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+
+public record DadosEndereco(
+        @NotBlank
+        String logradouro,
+
+        @NotBlank
+        String bairro,
+
+        @NotBlank
+        @Pattern(regexp = "\\d{8}") // Corrigido para validar 8 dígitos
+        String cep,
+
+        @NotBlank
+        String cidade,
+
+        @NotBlank
+        String uf,
+
+        String complemento,
+
+        String numero
+) {
+}
