+++ src/main/java/med/voll/api/domain/entities/medico/DadosListagemMedico.java
@@ -1,0 +1,9 @@
+package med.voll.api.domain.entities.medico;
+
+import med.voll.api.domain.enums.Especialidade;
+
+public record DadosListagemMedico(Long id, String nome, String email, String crm, Especialidade especialidade) {
+    public DadosListagemMedico(Medico medico){
+        this(medico.getId(), medico.getNome(), medico.getEmail(), medico.getCrm(), medico.getEspecialidade());
+    }
+}
