+++ src/main/java/med/voll/api/domain/entities/paciente/DadosAtualizacaoPaciente.java
@@ -1,0 +1,12 @@
+package med.voll.api.domain.entities.paciente;
+
+import jakarta.validation.constraints.NotNull;
+import med.voll.api.domain.entities.endereco.DadosEndereco;
+
+public record DadosAtualizacaoPaciente(
+        @NotNull
+        Long id,
+        String nome,
+        String telefone,
+        DadosEndereco endereco) {
+}
