+++ src/main/java/med/voll/api/domain/entities/paciente/DadosCadastroPaciente.java
@@ -1,0 +1,30 @@
+package med.voll.api.domain.entities.paciente;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
+import med.voll.api.domain.entities.endereco.DadosEndereco;
+
+public record DadosCadastroPaciente(
+        @NotBlank
+        String nome,
+
+        @NotBlank
+        @Email
+        String email,
+
+        @NotBlank
+        String telefone,
+
+        @NotBlank
+        @Pattern(regexp = "\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}")
+
+        String cpf,
+
+        @NotNull
+        @Valid
+        DadosEndereco endereco
+) {
+}
