+++ src/main/java/med/voll/api/domain/entities/paciente/DadosDetalhamentoPaciente.java
@@ -1,0 +1,9 @@
+package med.voll.api.domain.entities.paciente;
+
+import med.voll.api.domain.entities.endereco.Endereco;
+
+public record DadosDetalhamentoPaciente(Long id, String nome, String telefone, String cpf, Endereco endereco) {
+    public DadosDetalhamentoPaciente(Paciente paciente) {
+        this(paciente.getId(), paciente.getNome(), paciente.getTelefone(), paciente.getCpf(), paciente.getEndereco());
+    }
+}
