+++ src/main/java/med/voll/api/domain/entities/paciente/DadosListagemPaciente.java
@@ -1,0 +1,7 @@
+package med.voll.api.domain.entities.paciente;
+
+public record DadosListagemPaciente(Long id, String nome, String email, String cpf) {
+    public DadosListagemPaciente(Paciente paciente ){
+        this(paciente.getId(), paciente.getNome(), paciente.getEmail(), paciente.getCpf());
+    }
+}
