+++ src/main/java/med/voll/api/domain/entities/usuario/DadosAutenticacao.java
@@ -1,0 +1,4 @@
+package med.voll.api.domain.entities.usuario;
+
+public record DadosAutenticacao(String login, String senha) {
+}
