+++ src/main/java/med/voll/api/domain/enums/Especialidade.java
@@ -1,0 +1,8 @@
+package med.voll.api.domain.enums;
+
+public enum Especialidade {
+    ORTOPEDIA,
+    CARDIOLOGIA,
+    GINECOLOGIA,
+    DERMATOLOGIA
+}
