--- src/main/java/med/voll/api/domain/medico/DadosAtualizacaoMedico.java
@@ -1,15 +1,0 @@
-package med.voll.api.domain.medico;
-
-import jakarta.validation.constraints.NotNull;
-import med.voll.api.domain.endereco.DadosEndereco;
-
-public record DadosAtualizacaoMedico(
-        @NotNull
-        Long id,
-
-        String telefone,
-
-        String nome,
-
-        DadosEndereco endereco) {
-}
