--- src/main/java/med/voll/api/domain/medico/DadosDetalhamentoMedico.java
@@ -1,9 +1,0 @@
-package med.voll.api.domain.medico;
-
-import med.voll.api.domain.endereco.Endereco;
-
-public record DadosDetalhamentoMedico(Long id, String nome, String email, String crm ,String telefone, Especialidade especialidade, Endereco endereco) {
-    public DadosDetalhamentoMedico(Medico medico) {
-        this(medico.getId(), medico.getNome(), medico.getEmail(), medico.getCrm(),medico.getTelefone(), medico.getEspecialidade(), medico.getEndereco());
-    }
-}
