--- src/main/java/med/voll/api/domain/medico/DadosListagemMedico.java
@@ -1,7 +1,0 @@
-package med.voll.api.domain.medico;
-
-public record DadosListagemMedico(Long id, String nome, String email, String crm, Especialidade especialidade) {
-    public DadosListagemMedico(Medico medico){
-        this(medico.getId(), medico.getNome(), medico.getEmail(), medico.getCrm(), medico.getEspecialidade());
-    }
-}
