--- src/main/java/med/voll/api/domain/medico/Especialidade.java
@@ -1,8 +1,0 @@
-package med.voll.api.domain.medico;
-
-public enum Especialidade {
-    ORTOPEDIA,
-    CARDIOLOGIA,
-    GINECOLOGIA,
-    DERMATOLOGIA
-}
