--- src/main/java/med/voll/api/domain/medico/Medico.java
@@ -1,141 +1,0 @@
-package med.voll.api.domain.medico;
-
-import jakarta.persistence.*;
-import med.voll.api.domain.endereco.Endereco;
-
-import java.util.Objects;
-import java.util.Optional;
-
-@Entity
-@Table(name = "medicos")
-public class Medico {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String nome;
-    private String email;
-    private String telefone;
-    private String crm;
-
-    @Enumerated(EnumType.STRING)
-    private Especialidade especialidade;
-
-    @Embedded
-    private Endereco endereco;
-
-    private boolean ativo = true;
-
-    // Construtor padrão
-    public Medico() {
-    }
-
-    // Construtor com todos os argumentos
-    public Medico(Long id, String nome, String email, String telefone, String crm, Especialidade especialidade, Endereco endereco, boolean ativo) {
-        this.id = id;
-        this.nome = nome;
-        this.email = email;
-        this.telefone = telefone;
-        this.crm = crm;
-        this.especialidade = especialidade;
-        this.endereco = endereco;
-        this.ativo = ativo; //
-    }
-
-    // Construtor com DadosCadastroMedico
-    public Medico(DadosCadastroMedico dados) {
-        this.ativo = true;
-        this.nome = dados.nome();
-        this.email = dados.email();
-        this.telefone = dados.telefone();
-        this.crm = dados.crm();
-        this.especialidade = Especialidade.valueOf(dados.especialidade());
-        this.endereco = new Endereco(dados.endereco());
-    }
-
-    // Getters e Setters
-    public Long getId() {
-        return id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public String getCrm() {
-        return crm;
-    }
-
-    public void setCrm(String crm) {
-        this.crm = crm;
-    }
-
-    public Especialidade getEspecialidade() {
-        return especialidade;
-    }
-
-    public void setEspecialidade(Especialidade especialidade) {
-        this.especialidade = especialidade;
-    }
-
-    public Endereco getEndereco() {
-        return endereco;
-    }
-
-    public void setEndereco(Endereco endereco) {
-        this.endereco = endereco;
-    }
-
-    public boolean isAtivo() {
-        return ativo;
-    }
-
-    public void setAtivo(boolean ativo) {
-        this.ativo = ativo;
-    }
-
-    // Equals e HashCode usando apenas o ID
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Medico medico = (Medico) o;
-        return Objects.equals(id, medico.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-
-    public void atualizarInformacoes(DadosAtualizacaoMedico dados) {
-        Optional.ofNullable(dados.nome()).ifPresent(valor -> this.nome = valor);
-        Optional.ofNullable(dados.telefone()).ifPresent(valor -> this.telefone = valor);
-        Optional.ofNullable(dados.endereco()).ifPresent(this.endereco::atualizarInformacoes);
-    }
-
-    public void desativar() {
-        this.ativo = false;
-    }
-}
