--- src/main/java/med/voll/api/domain/medico/MedicoRepository.java
@@ -1,12 +1,0 @@
-package med.voll.api.domain.medico;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-
-
-public interface MedicoRepository extends JpaRepository<Medico, Long> {
-    Page<Medico> findAllByAtivoTrue(Pageable paginacao);
-}
