--- src/main/java/med/voll/api/domain/paciente/DadosAtualizacaoPaciente.java
@@ -1,12 +1,0 @@
-package med.voll.api.domain.paciente;
-
-import jakarta.validation.constraints.NotNull;
-import med.voll.api.domain.endereco.DadosEndereco;
-
-public record DadosAtualizacaoPaciente(
-        @NotNull
-        Long id,
-        String nome,
-        String telefone,
-        DadosEndereco endereco) {
-}
