--- src/main/java/med/voll/api/domain/paciente/DadosDetalhamentoPaciente.java
@@ -1,9 +1,0 @@
-package med.voll.api.domain.paciente;
-
-import med.voll.api.domain.endereco.Endereco;
-
-public record DadosDetalhamentoPaciente(Long id, String nome, String telefone, String cpf, Endereco endereco) {
-    public DadosDetalhamentoPaciente(Paciente paciente) {
-        this(paciente.getId(), paciente.getNome(), paciente.getTelefone(), paciente.getCpf(), paciente.getEndereco());
-    }
-}
