--- src/main/java/med/voll/api/domain/paciente/DadosListagemPaciente.java
@@ -1,7 +1,0 @@
-package med.voll.api.domain.paciente;
-
-public record DadosListagemPaciente(Long id, String nome, String email, String cpf) {
-    public DadosListagemPaciente(Paciente paciente ){
-        this(paciente.getId(), paciente.getNome(), paciente.getEmail(), paciente.getCpf());
-    }
-}
