--- src/main/java/med/voll/api/domain/paciente/Paciente.java
@@ -1,132 +1,0 @@
-package med.voll.api.domain.paciente;
-
-import jakarta.persistence.*;
-import med.voll.api.domain.endereco.Endereco;
-
-import java.util.Objects;
-import java.util.Optional;
-
-@Table(name = "pacientes")
-@Entity(name = "Paciente")
-public class Paciente {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String nome;
-    private String email;
-    private String telefone;
-    private String cpf;
-
-    @Embedded
-    private Endereco endereco;
-
-    private Boolean ativo;
-
-    // Construtor padrão
-    public Paciente() {
-    }
-
-    // Construtor com todos os argumentos
-    public Paciente(Long id, String nome, String email, String telefone, String cpf, Endereco endereco, Boolean ativo) {
-        this.id = id;
-        this.nome = nome;
-        this.email = email;
-        this.telefone = telefone;
-        this.cpf = cpf;
-        this.endereco = endereco;
-        this.ativo = ativo;
-    }
-
-    // Construtor com DadosCadastroPaciente
-    public Paciente(DadosCadastroPaciente dados) {
-        this.ativo = true;
-        this.nome = dados.nome();
-        this.email = dados.email();
-        this.telefone = dados.telefone();
-        this.cpf = dados.cpf();
-        this.endereco = new Endereco(dados.endereco());
-    }
-
-    // Métodos Getters e Setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public Endereco getEndereco() {
-        return endereco;
-    }
-
-    public void setEndereco(Endereco endereco) {
-        this.endereco = endereco;
-    }
-
-    public Boolean getAtivo() {
-        return ativo;
-    }
-
-    public void setAtivo(Boolean ativo) {
-        this.ativo = ativo;
-    }
-
-    // Método para atualizar informações com Optional
-    public void atualizarInformacoes(DadosAtualizacaoPaciente dados) {
-        Optional.ofNullable(dados.nome()).ifPresent(valor -> this.nome = valor);
-        Optional.ofNullable(dados.telefone()).ifPresent(valor -> this.telefone = valor);
-        Optional.ofNullable(dados.endereco()).ifPresent(this.endereco::atualizarInformacoes);
-    }
-
-    public void excluir() {
-        this.ativo = false;
-    }
-
-    // Equals e HashCode usando apenas o ID
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Paciente paciente = (Paciente) o;
-        return Objects.equals(id, paciente.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
