--- src/main/java/med/voll/api/domain/paciente/PacienteRepository.java
@@ -1,9 +1,0 @@
-package med.voll.api.domain.paciente;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PacienteRepository extends JpaRepository <Paciente, Long> {
-    Page<Paciente> findAllByAtivoTrue(Pageable paginacao);
- }
