--- src/main/java/med/voll/api/domain/usuario/DadosAutenticacao.java
@@ -1,4 +1,0 @@
-package med.voll.api.domain.usuario;
-
-public record DadosAutenticacao(String login, String senha) {
-}
