--- src/main/java/med/voll/api/domain/usuario/UsuarioRepository.java
@@ -1,9 +1,0 @@
-package med.voll.api.domain.usuario;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
-    Optional<Usuario> findByLogin(String login);
-}
