+++ src/main/java/med/voll/api/repositories/MedicoRepository.java
@@ -1,0 +1,10 @@
+package med.voll.api.repositories;
+
+import med.voll.api.domain.entities.medico.Medico;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MedicoRepository extends JpaRepository<Medico, Long> {
+    Page<Medico> findAllByAtivoTrue(Pageable paginacao);
+}
