+++ src/main/java/med/voll/api/repositories/UsuarioRepository.java
@@ -1,0 +1,10 @@
+package med.voll.api.repositories;
+
+import med.voll.api.domain.entities.usuario.Usuario;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+    Optional<Usuario> findByLogin(String login);
+}
