+++ src/main/resources/application.properties
@@ -3,10 +3,7 @@ spring.datasource.username=root
 spring.datasource.password=123asd456zxC**
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
-
-
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
-spring.flyway.repair=true
 
 server.error.include-stacktrace=never
